Reroute some path, made a home path ('/home') and ('/recipes') now is list of recipes
